classroom/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.30.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
bcrypt>=4.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
python-magic==0.4.27
redis==5.1.1
celery==5.4.0
slowapi==0.1.9
httpx==0.27.2
openpyxl==3.1.5
python-dotenv==1.0.1
email-validator==2.2.0
bleach==6.1.0
tinycss2==1.4.0
aiofiles==24.1.0
pytest==8.3.3
